requirements.txt
crawl4ai>=0.6.0
firecrawl-py>=1.0.0
anthropic>=0.40.0
groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
rich>=13.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Live server
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9

# OSINT
sherlock-project>=0.15.0
